requirements.txt
atlassian-python-api==3.41.0
apscheduler==3.10.4
jinja2==3.1.2
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
